THIS BETTER WORK

+++ index.js
@@ -104,13 +104,14 @@ function runCmd(msg) {
     } else if (clean > -1) {
         cleanChannel(msg);
     } else if (reset > -1) {
-        data.mods.forEach(user => {
+        if (data.mods.forEach(user => {
             if (msg.author.id == user.id) {
                 deleteAll(msg);
-                return;
+                return true;
             }
-        });
-        msg.channel.send("**Hey! you're not a mod**").then(botMsg => { deleteMsg(botMsg, 5000); }).catch();
+        }) != true) {
+            msg.channel.send("**Hey! you're not a mod**").then(botMsg => { deleteMsg(botMsg, 5000); }).catch();
+        }
     } else {
         return;
     }
